Added random-bot script

+++ commons/include/ir_operands.h
@@ -27,7 +27,7 @@ enum ir_opd_type_t {
 union ir_opd_value_t {
     reg_t       reg;
     int32_t     offset;
-    int32_t     imm;
+    int64_t     imm;
     const char* global_label_name;
     const char* string_literal;
     size_t      local_label_number;
